update some parts to the changes
