+++ src/main/java/de/mari_023/fabric/ae2wtlib/client/ae2wtlibclient.java
@@ -1,6 +1,6 @@
 package de.mari_023.fabric.ae2wtlib.client;
 
-import appeng.container.implementations.WirelessCraftConfirmContainer;
+import appeng.container.me.crafting.WirelessCraftConfirmContainer;
 import appeng.container.implementations.WirelessCraftingStatusContainer;
 import appeng.util.item.AEItemStack;
 import de.mari_023.fabric.ae2wtlib.Config;
