+++ src/main/java/de/mari_023/fabric/ae2wtlib/util/WirelessCraftConfirmScreen.java
@@ -5,7 +5,7 @@ import appeng.api.storage.data.IAEItemStack;
 import appeng.api.storage.data.IItemList;
 import appeng.client.gui.AEBaseScreen;
 import appeng.client.gui.widgets.Scrollbar;
-import appeng.container.implementations.WirelessCraftConfirmContainer;
+import appeng.container.me.crafting.WirelessCraftConfirmContainer;
 import appeng.core.Api;
 import appeng.core.localization.GuiText;
 import appeng.core.sync.network.NetworkHandler;
